Divided aperture.py into package with individual files for each aperture type.
