+++ pygerber/meta/aperture/__init__.py
@@ -1,0 +1,6 @@
+from .aperture import Aperture
+from .circular import CircularAperture
+from .rectangular import RectangularAperture
+from .polygon import PolygonAperture
+from .region import RegionApertureManager
+from .custom import CustomAperture
