+++ pygerber/meta/aperture/circular.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import List, Tuple
+
+from pygerber.mathclasses import BoundingBox
+from pygerber.meta.arc_util_mixin import ArcUtilMixin
+from pygerber.meta.spec import ArcSpec, FlashSpec, LineSpec, Spec
+from pygerber.tokens.add import ADD_Token
+from .aperture import Aperture
+
+
+class CircularAperture(Aperture):
+
+    DIAMETER: float
+    HOLE_DIAMETER: float
+
+    def __init__(self, args: ADD_Token.ARGS, broker) -> None:
+        self.broker = broker
+        self.HOLE_DIAMETER = args.HOLE_DIAMETER
+        self.DIAMETER = args.DIAMETER
+
+    def bbox(self) -> BoundingBox:
+        d_half = self.DIAMETER / 2
+        return BoundingBox(
+            -d_half,
+            d_half,
+            d_half,
+            -d_half,
+        )
